+++ app/components/tooltip-on-element-lazy.js
@@ -1,0 +1,3 @@
+export { default } from 'ember-tooltips/components/tooltip-on-element-lazy';
+
+// TODO(Andrew) find a better name for this
